Moved util files to utils
