+++ requirements.txt
@@ -1,15 +1,9 @@
-torch>=1.1
-batchgenerators
-torchvision
+matplotlib
+tensorflow==2.5
+keras==2.4
 numpy
-tqdm
-tensorboard>=1.14
 sklearn
 pandas
-torchsummary
-xlrd
-openpyxl
-optuna
-matplotlib
-git+git://github.com/MonsieurWave/torchio#egg=torchio
-git+git://github.com/shijianjian/EfficientNet-PyTorch-3D#egg=efficientnet_pytorch_3d
+monai
+scikit-image
+gputil
