--- scope/requirements.txt
@@ -1,9 +1,0 @@
-matplotlib
-tensorflow==2.5
-keras==2.4
-numpy
-sklearn
-pandas
-monai
-skimage
-gputil
